+++ src/logic/EditHandler.java
@@ -7,6 +7,7 @@ import java.util.Arrays;
 import java.util.Collections;
 
 import application.Task;
+import application.TaskCreator;
 import parser.IndexParser;
 import application.TaskComparator;
